removed some files and fixed several typos

+++ src/linkedcells/CutoffDistances.jl
@@ -6,8 +6,8 @@ struct CutoffDistances
 
   # To store all distances smaller than the cutoff
   d :: Vector{Float64} # All distances smaller than the cutoff
-  iat :: Vector{Float64} # atom of the solute
-  jat :: Vector{Float64} # atom of the solvent
+  iat :: Vector{Int64} # atom of the solute
+  jat :: Vector{Int64} # atom of the solvent
 
 end
